+++ umbra/components/addons/traceUi/nodes.py
@@ -129,9 +129,19 @@ class ModuleNode(umbra.ui.nodes.GraphModelNode):
 		:param attributesFlags: Attributes flags. ( Integer )
 		"""
 
+		self.synchronizeAttributes()
+
+	def synchronizeAttributes(self, attributesFlags=int(Qt.ItemIsSelectable | Qt.ItemIsEnabled)):
+		"""
+		This method synchronizes the node attributes.
+		
+		:param attributesFlags: Attributes flags. ( Integer )
+		:return: Method success. ( Boolean )
+		"""
+
 		isTraced = foundations.trace.isTraced(self.__module)
 		self["traced"] = umbra.ui.nodes.GraphModelAttribute(name="traced",
 															value=isTraced,
 															roles=dict.fromkeys((Qt.DisplayRole, Qt.EditRole),
 																foundations.strings.encode(isTraced).title()),
-															flags=int(Qt.ItemIsSelectable | Qt.ItemIsEnabled))
+															flags=attributesFlags)